% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dates_differences.R
\name{add_dates_differences}
\alias{add_dates_differences}
\title{Adding dates difference}
\usage{
add_dates_differences(df)
}
\arguments{
\item{df}{a data frame of the structure of PACS data. It should contain at
least 4 Date columns: "onset", "hospitalization", "consultation" and
"sample_collection".}
}
\value{
the data frame df augmented with 6 columns.
}
\description{
This function adds 6 dates difference between the 4 following dates variables:
"onset", "hospitalization", "consultation" and "sample_collection". The names
of these new variables all start with "diff_": "diff_onset_hospitalization",
"diff_onset_consultation", etc...
}
\examples{
pacs <- readr::read_csv("../../R code/pacs/data/pacs.csv",
                        col_types = paste(c("icfnD", rep("c", 5), rep("D", 4), rep("f", 3)), collapse = ""))
require(magrittr)
pacs \%<>\% add_dates_differences()

}
\author{
Marc Choisy
}
