% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_tests.R
\name{add_tests}
\alias{add_tests}
\title{Adding test information to the PACS data frame}
\usage{
add_tests(df)
}
\arguments{
\item{df}{a data frame of the structure of PACS data. It should contain at
least 3 columns: "pcr" and "ns1" for which positive and negative should be
coded "positive" and "negative" respectively and "serotype" where dengue
serotype x should be coded by a character string containing "dengue", with all
other values not containing this pattern.}
}
\description{
This funtion adds 2 columns to a data frame of PACS structure: one column
named "tested" informing whether a test (either PCR, NS1 or serotype) has
been performed on the sample and a column named "confirmed" indicating whether
the sample is confirmed positive.
}
\examples{
pacs <- readr::read_csv("../../R code/pacs/data/pacs.csv",
                        col_types = paste(c("icfnD", rep("c", 5), rep("D", 4), rep("f", 3)), collapse = ""))
require(magrittr)
pacs \%<>\% adds_test()

}
\author{
Marc Choisy
}
