% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_onset.R
\name{correct_onset}
\alias{correct_onset}
\title{Generates corrected onset dates}
\usage{
correct_onset(df)
}
\arguments{
\item{df}{a data frame of the structure of PACS data. It should contain at
least 4 Date columns: "onset", "hospitalization", "consultation" and
"sample_collection".}
}
\value{
a vector of dates.
}
\description{
This generates corrected onset dates from a data frame that contains 4 dates
columns: "onset", "hospitalization", "consultation", "sample_collection".
}
\examples{
require(dplyr)

pacs <- readr::read_csv("../../R code/pacs/data/pacs.csv",
                        col_types = paste(c("icfnD", rep("c", 5), rep("D", 4), rep("f", 3)), collapse = ""))

pacs \%>\%
  mutate(onset2 = correct_onset(.))

}
\author{
Marc Choisy
}
