% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_at_any.R
\name{filter_at_any}
\alias{filter_at_any}
\title{Filter at at least}
\usage{
filter_at_any(df, pattern, f, sel)
}
\arguments{
\item{df}{a data frame of the structure of PACS data. It should contain at
least 4 Date columns: "onset", "hospitalization", "consultation" and
"sample_collection".}
}
\description{
This function filters a data frame so that at least one of the columns with a
name matching a given pattern fullfills a condition specified by a given
function.
}
\examples{
require(dplyr)

pacs <- readr::read_csv("../../R code/pacs/data/pacs.csv",
                        col_types = paste(c("icfnD", rep("c", 5), rep("D", 4), rep("f", 3)), collapse = ""))

pacs \%>\%
  ecomore::add_dates_differences() \%>\%
  filter_at_any("diff_", function(x) abs(x) > 15, starts_with) \%>\%
  select(id, onset, hospitalization, consultation, sample_collection)

}
\author{
Marc Choisy
}
